+++ src/ErrorState/index.ts
@@ -1,2 +1,2 @@
 export { ErrorState, default } from './ErrorState';
-export type { ErrorStateProps } from './ErrorState';
+export type { ErrorStateAction, ErrorStateProps } from './ErrorState';